move concentric to concentric2d
