+++ src/layout/concentric2d.ts
@@ -1,145 +1,62 @@
-import { LayoutFactoryProps, LayoutStrategy } from './types';
+import { LayoutFactoryProps } from './types';
 import { buildNodeEdges } from './layoutUtils';
-import { InternalGraphNode, InternalGraphPosition } from '../types';
-import Graph from 'graphology';
 
 export interface ConcentricLayoutInputs extends LayoutFactoryProps {
   /**
-   * Radius of the innermost circle.
+   * Base radius of the innermost circle.
    */
-  minRadius?: number;
+  radius: number;
   /**
-   * Radius of the outermost circle.
+   * Distance between circles.
    */
-  maxRadius?: number;
-  /**
-   * Center point of the layout.
-   */
-  center?: [number, number];
-  /**
-   * Whether to use degree centrality for positioning.
-   */
-  useDegreeCentrality?: boolean;
-  /**
-   * Spacing between rings.
-   */
-  spacingFactor?: number;
-  /**
-   * Whether to start angle from the top (12 o'clock position).
-   */
-  startAngle?: number;
+  concentricSpacing?: number;
 }
 
-export function concentric2d({
+export function concentricLayout({
   graph,
+  radius = 40,
   drags,
-  minRadius = 50,
-  maxRadius = 300,
-  center = [0, 0],
-  useDegreeCentrality = true,
-  spacingFactor = 1.25,
-  startAngle = 0,
-  getNodePosition
-}: ConcentricLayoutInputs): LayoutStrategy {
+  getNodePosition,
+  concentricSpacing = 100
+}: ConcentricLayoutInputs) {
   const { nodes, edges } = buildNodeEdges(graph);
 
-  if (nodes.length === 0) {
-    return {
-      step() {
-        return true;
-      },
-      getNodePosition() {
-        return {
-          x: 0,
-          y: 0,
-          z: 0,
-          id: '',
-          data: {},
-          links: [],
-          index: 0,
-          vx: 0,
-          vy: 0
-        };
-      }
+  const metrics = nodes.map(node => ({
+    id: node.id,
+    metric: graph.degree(node.id)
+  }));
+
+  // Sort by importance (degree)
+  metrics.sort((a, b) => b.metric - a.metric);
+
+  const layout: Record<string, { x: number; y: number }> = {};
+  let level = 0;
+  let indexInLevel = 0;
+
+  const getNodesInLevel = (level: number) => {
+    const circumference = 2 * Math.PI * (radius + level * concentricSpacing);
+    const minNodeSpacing = 40; // Minimum spacing between nodes in px
+    return Math.floor(circumference / minNodeSpacing);
+  };
+
+  for (let i = 0; i < metrics.length; i++) {
+    const nodesInCurrentLevel = getNodesInLevel(level);
+
+    if (indexInLevel >= nodesInCurrentLevel) {
+      level++;
+      indexInLevel = 0;
+    }
+
+    const r = radius + level * concentricSpacing;
+    const angle = (2 * Math.PI * indexInLevel) / nodesInCurrentLevel;
+
+    layout[metrics[i].id] = {
+      x: r * Math.cos(angle),
+      y: r * Math.sin(angle)
     };
+
+    indexInLevel++;
   }
-
-  // Calculate degree centrality for each node
-  const nodeDegrees = new Map<string, number>();
-  nodes.forEach(node => {
-    const degree = graph.degree(node.id);
-    nodeDegrees.set(node.id, degree);
-  });
-
-  // Sort nodes by degree (descending) for centrality-based positioning
-  const sortedNodes = [...nodes].sort((a, b) => {
-    const degreeA = nodeDegrees.get(a.id) || 0;
-    const degreeB = nodeDegrees.get(b.id) || 0;
-    return degreeB - degreeA;
-  });
-
-  // Determine number of rings and assign nodes to rings
-  const rings: InternalGraphNode[][] = [];
-  const degreeValues = Array.from(nodeDegrees.values());
-  const maxDegree = Math.max(...degreeValues);
-  const minDegree = Math.min(...degreeValues);
-
-  if (useDegreeCentrality && maxDegree > minDegree) {
-    // Create rings based on degree centrality
-    const degreeRange = maxDegree - minDegree;
-    const numRings = Math.min(5, degreeRange + 1);
-
-    // Initialize rings
-    for (let i = 0; i < numRings; i++) {
-      rings[i] = [];
-    }
-
-    // Assign nodes to rings based on degree
-    sortedNodes.forEach(node => {
-      const degree = nodeDegrees.get(node.id) || 0;
-      const ringIndex = Math.floor(
-        ((degree - minDegree) / degreeRange) * (numRings - 1)
-      );
-      rings[ringIndex].push(node);
-    });
-  } else {
-    // Fallback: distribute nodes evenly across rings
-    const numRings = Math.min(5, Math.ceil(Math.sqrt(nodes.length)));
-    const nodesPerRing = Math.ceil(nodes.length / numRings);
-
-    for (let i = 0; i < numRings; i++) {
-      rings[i] = sortedNodes.slice(i * nodesPerRing, (i + 1) * nodesPerRing);
-    }
-  }
-
-  // Filter out empty rings
-  const nonEmptyRings = rings.filter(ring => ring.length > 0);
-
-  // Calculate positions for each node
-  const nodePositions = new Map<string, InternalGraphPosition>();
-  const [centerX, centerY] = center;
-
-  nonEmptyRings.forEach((ringNodes, ringIndex) => {
-    // Calculate radius with spacing factor
-    const radius =
-      minRadius +
-      (ringIndex * spacingFactor * (maxRadius - minRadius)) /
-        (nonEmptyRings.length - 1);
-
-    // Distribute nodes evenly around the circle
-    ringNodes.forEach((node, nodeIndex) => {
-      const angle = startAngle + (2 * Math.PI * nodeIndex) / ringNodes.length;
-      const x = centerX + radius * Math.cos(angle);
-      const y = centerY + radius * Math.sin(angle);
-
-      nodePositions.set(node.id, {
-        ...node.position,
-        x,
-        y,
-        z: 0
-      });
-    });
-  });
 
   return {
     step() {
@@ -148,17 +65,14 @@ export function concentric2d({
     getNodePosition(id: string) {
       if (getNodePosition) {
         const pos = getNodePosition(id, { graph, drags, nodes, edges });
-        if (pos) {
-          return pos;
-        }
+        if (pos) return pos;
       }
 
       if (drags?.[id]?.position) {
-        // If we dragged, we need to use that position
-        return drags?.[id]?.position as any;
+        return drags[id].position as any;
       }
 
-      return nodePositions.get(id);
+      return layout[id];
     }
   };
 }
